SelectorBase class abstraction with core selector functionality to extend upon
