+++ frontend/src/app/components/selectors/RectROISelector.ts
@@ -1,9 +1,10 @@
 import {ROISelection2D} from '../../model/ROISelection2D';
 import {Selector} from './Selector';
 import {EventEmitter} from "@angular/core";
+import {SelectorBase} from "./SelectorBase";
 import {SliceSelection} from "../../model/SliceSelection";
 
-export class RectROISelector implements Selector<ROISelection2D> {
+export class RectROISelector extends SelectorBase<ROISelection2D> {
 	readonly STYLE = {
 		SELECTION_FONT_SIZE: 14,
 		SELECTION_LINE_DENSITY: [6],
@@ -12,18 +13,8 @@ export class RectROISelector implements Selector<ROISelection2D> {
 		ARCHIVED_SELECTION_COLOR: '#5f27e5'
 	};
 
-	selectedArea: ROISelection2D;
-	selections: Map<number, ROISelection2D>;
-	archivedSelections: Array<ROISelection2D>;
-	canvasCtx: CanvasRenderingContext2D;
-	protected currentSlice;
-	protected mouseDrag = false;
-	protected canvasPosition: ClientRect;
-	canvasSize: { width: number, height: number };
-
-	public stateChange: EventEmitter<number>;
-
 	constructor(canvas: HTMLCanvasElement) {
+		super();
 		this.canvasCtx = canvas.getContext('2d');
 		this.canvasSize = {
 			width: canvas.width,
@@ -34,14 +25,6 @@ export class RectROISelector implements Selector<ROISelection2D> {
 		this.selectedArea = undefined;
 		this.currentSlice = undefined;
 		this.stateChange = new EventEmitter<number>();
-	}
-
-	public getStateChangeEmitter(): EventEmitter<number> {
-		return this.stateChange;
-	}
-
-	public getSelections(): ROISelection2D[] {
-		return Array.from(this.selections.values());
 	}
 
 	public formArchivedSelections(selectionMap: ROISelection2D[]): ROISelection2D[] {
@@ -134,125 +117,7 @@ export class RectROISelector implements Selector<ROISelection2D> {
 		}
 	}
 
-	public clearCanvasSelection(): void {
-		console.log("RectROISelector | clearCanvasSelection");
-		this.canvasCtx.clearRect(0, 0, this.canvasSize.width, this.canvasSize.height);
-		this.drawPreviousSelections();
-	}
-
-	public clearData(): void {
-		this.selectedArea = undefined;
-		this.selections = new Map<number, ROISelection2D>();
-		this.archivedSelections = [];
-		this.stateChange = new EventEmitter<number>();
-		this.clearCanvasSelection();
-	}
-
-	public addCurrentSelection(): void {
-		if (this.selectedArea) {
-			console.log("RectROISelector | addCurrentSelection");
-			this.selections.set(this.currentSlice, this.selectedArea);
-			this.stateChange.emit(this.currentSlice);
-			this.clearSelectedArea();
-		}
-	}
-
-	private clearSelectedArea() {
-		this.selectedArea = undefined;
-	}
-
-	public updateCurrentSlice(currentSliceId: number): void {
-		this.currentSlice = currentSliceId;
-	}
-
-	public updateCanvasPosition(canvasRect: ClientRect) {
-		this.canvasPosition = canvasRect;
-	}
-
-	public hasArchivedSelections(): boolean {
-		return this.archivedSelections.length > 0;
-	}
-
-	public hasValidSelection(): boolean {
-		return this.selections.size >= 1;
-	}
-
-	public hasSliceSelection(): boolean {
-		return !!this.selections.get(this.currentSlice) || !!this.selectedArea;
-	}
-
-	public archiveSelections(selectionMap?: Array<ROISelection2D>): void {
-		if (!selectionMap) {
-			selectionMap = Array.from(this.selections.values());
-		}
-		selectionMap.forEach((value: ROISelection2D) => {
-			this.archivedSelections.push(value);
-		});
-	}
-
-	public removeCurrentSelection(): void {
-		if (this.hasSliceSelection()) {
-			this.selections.delete(this.currentSlice);
-			this.selectedArea = undefined;
-
-			this.clearCanvasSelection();
-		}
-	}
-
-	public clearSelections() {
-		this.selections.clear();
-	}
-
-	public updateCanvasHeight(height: number): void {
-		if (height && height > 0) {
-			this.canvasSize.height = height;
-		} else {
-			console.warn('RectROISelector | updateCanvasHeight - non numeric/positive height provided', height);
-		}
-	}
-
-	public updateCanvasWidth(width: number): void {
-		if (width && width > 0) {
-			this.canvasSize.width = width;
-		} else {
-			console.warn('RectROISelector | updateCanvasWidth - non numeric/positive width provided', width);
-		}
-	}
-
-	public removeSelection(selectionId: number): void {
-		if(selectionId == this.currentSlice) {
-			this.removeCurrentSelection();
-		} else {
-			this.selections.delete(selectionId);
-			this.clearCanvasSelection();
-		}
-	}
-
-	public hideSelection(selectionId: number, newValue: boolean): void {
-		let selection: SliceSelection = this.selections.get(selectionId);
-		if(selection) {
-			if (selection.hidden != newValue) {
-				selection.hidden = newValue;
-				this.redrawSelections();
-			}
-		} else {
-			console.warn('RectROISelector | hideSelection | cannot hide non present selection!');
-		}
-	}
-
-	public pinSelection(selectionId: number, newValue: boolean): void {
-		let selection: SliceSelection = this.selections.get(selectionId);
-		if(selection) {
-			if (selection.pinned != newValue) {
-				selection.pinned = newValue;
-				this.redrawSelections();
-			}
-		} else {
-			console.warn('RectROISelector | hideSelection | cannot pin non present selection!');
-		}
-	}
-
-	private redrawSelections(): void {
+	public redrawSelections(): void {
 		this.clearCanvasSelection();
 
 		this.drawPreviousSelections();
